+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - [fix: fix ConfigChangeListener ut bug](https://github.com/Tencent/spring-cloud-tencent/pull/1663)
 - [feat: support shortest response time lb and least connection lb](https://github.com/Tencent/spring-cloud-tencent/pull/1637)
 - [feat:support traffic mirroring.](https://github.com/Tencent/spring-cloud-tencent/pull/1647)
+- [feat: add delay test interface in quickstart.](https://github.com/Tencent/spring-cloud-tencent/pull/1666)